#8.4 Protecting Video Routes

+++ controller/videoController.js
@@ -68,13 +68,19 @@ export const detailVideo = async (req, res) => {
 };
 
 /* 동영상 세부정보 편집 */
-export const geteditVideo = async (req, res) => {
+export const getEditVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
+
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -98,8 +104,15 @@ export const deleteVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
+
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
